move Books and Persons into separate classes and initialize in main
